+++ glean/src/core/error/index.ts
@@ -59,14 +59,14 @@ export default class ErrorManager {
    *        prepended to the message.
    * @param numErrors The number of errors of the same type to report.
    */
-  async record (
+  async record(
     metric: MetricType,
     error: ErrorType,
-    message: string,
+    message: unknown,
     numErrors = 1
   ): Promise<void> {
     const errorMetric = getErrorMetricForMetric(metric, error);
-    log(createLogTag(metric), `${metric.baseIdentifier()}: ${message}`);
+    log(createLogTag(metric), [`${metric.baseIdentifier()}:`, message]);
     if (numErrors > 0) {
       await CounterMetricType._private_addUndispatched(errorMetric, numErrors);
     } else {